+++ src/getStepStartStates.test.ts
@@ -95,6 +95,7 @@ import {
 } from './steps/resource-manager/event-grid';
 import { AdvisorSteps } from './steps/resource-manager/advisor';
 import { SecuritySteps } from './steps/resource-manager/security';
+import { PolicySteps } from './steps/resource-manager/policy/constants';
 
 describe('getStepStartStates', () => {
   test('all steps represented', () => {
@@ -172,6 +173,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_EVENT_GRID_DOMAIN_TOPIC_SUBSCRIPTIONS]: { disabled: true },
       [AdvisorSteps.RECOMMENDATIONS]: { disabled: true },
       [SecuritySteps.ASSESSMENTS]: { disabled: true },
+      [PolicySteps.POLICY_ASSIGNMENTS]: { disabled: true },
     });
   });
 
@@ -245,6 +247,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_EVENT_GRID_DOMAIN_TOPIC_SUBSCRIPTIONS]: { disabled: true },
       [AdvisorSteps.RECOMMENDATIONS]: { disabled: true },
       [SecuritySteps.ASSESSMENTS]: { disabled: true },
+      [PolicySteps.POLICY_ASSIGNMENTS]: { disabled: true },
     });
   });
 
@@ -318,6 +321,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_EVENT_GRID_DOMAIN_TOPIC_SUBSCRIPTIONS]: { disabled: false },
       [AdvisorSteps.RECOMMENDATIONS]: { disabled: false },
       [SecuritySteps.ASSESSMENTS]: { disabled: false },
+      [PolicySteps.POLICY_ASSIGNMENTS]: { disabled: false },
     });
   });
 });