+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Added `azure_policy_assignment` entities
+- Added `ANY_SCOPE|has|azure_policy_assignment` relationships. These can target
+  any scoped entity within Azure.
+
 ## 5.5.2 - 2020-10-29
 
 ### Changed